# Avoid extra ReadAt calls when seeking pages (sync mode)