@stoewer 2025-09-05
> I looked into the flaky async test: to me it really looks like the failure is caused by asyncPages and not by the underlying `FilePages`

Yeah, I was wrong about that: the flakiness was caused by by `FilePages.ReadPage` in the branch of the code that serves the last returned page. I've fixed it now and also added some small improvements to `TestSeekToRowGeneral`